+++ tilebox-workflows/tests/proto/__init__.py
@@ -1,0 +1,2 @@
+# to re-generate proto files, run:
+# > protoc --python_out=pyi_out:. test.proto