+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.0] - 2024-09-30
+
 ### Added
 
 - Create collection RPC stub and service method
 - Ingest and delete datapoints RPC stubs and service methods
 - Configure a default console logger if `get_logger()` is called without any configured loggers.
+- Callback function for reporting progress percentage for timeseries datasets `.load()` method.
+
+### Changed
+
+- Updated TimeInterval and TimeChunk tasks to use protobuf messages from datasets package.
 
 ## [0.25.1] - 2024-09-12
 
@@ -28,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.25.1...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/tilebox/tilebox-python/compare/v0.25.1...v0.26.0
 [0.25.1]: https://github.com/tilebox/tilebox-python/compare/v0.25.0...v0.25.1
 [0.25.0]: https://github.com/tilebox/tilebox-python/releases/tag/v0.25.0